and one more

+++ restful-web-services/src/test/java/com/sept/rest/webservices/restfulwebservices/UserBean/UserBeanControllerTest.java
@@ -27,7 +27,7 @@ public class UserBeanDataBaseTest {
     @MockBean
     private UserBeanService userBeanService;
 
-    UserBean mockUser = new UserBean((long) 1, "Sherry1377", "Shahrzad", "Rafezi", "Password123!", false);
+    UserBean mockUser = new UserBean(1, "Sherry1377", "Shahrzad", "Rafezi", "Password123!", false);
 
     @Test
     public void retrieveDetailsForUser() throws Exception {
